Log an error & recover in case of corrupted pre-serialized key/value
The goal is to be able to detect memory corruption on the client side
which would result in sending a message with valid CRC but invalid format.

When SerializeOnProduce is enabled, key & value are serialized when
entering the driver and then freed. As we no longer have them, the only
case we can detect is if the length is too small (it cannot be empty as it
contains at least the key & value sizes).

As serializing such message would result in sending a produce request
which would be accepted by the broker but would break the consumer, we
recover by simulating a null key & value.

Change-Id: I6ec9d9dc5b8b7cec3b5c667fc4442be810f72754
JIRA: BES-13701